requirements.txt
Flask==3.0.0
numpy==1.24.3
pandas==1.5.3
plotly==5.18.0
configparser==6.0.0
python-dotenv==1.0.0
paho-mqtt==1.6.1
flask-socketio==5.3.6
eventlet==0.33.3
pyyaml>=6.0.1
sqlite3
python-dateutil 